create Matrix.h and Matrix.cpp
